Split HistoryScreen into per-view files
HistoryScreen.kt had grown to ~665 lines with nine composables across
calendar, trends, and permission concerns. Splits it into four
sibling files so each one fits a single mental scope:

- HistoryScreen.kt: top-level composable, tab routing, pager
- HistoryCalendarView.kt: CalendarView + DayDetailCard
- HistoryTrendsView.kt: TrendsView + LegendSwatch + TrendChart
- HistoryPermissions.kt: ReadPermissionRequired

SystolicColor and DiastolicColor drop the file-private modifier so
the calendar can keep its "has readings" dot coloured the same as
the diastolic series. DailyAverage stays private to the trends file
since it has no other consumer. No behaviour changes.
